% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bootstrap_function.R
\name{bootstrap_logistic_confidence}
\alias{bootstrap_logistic_confidence}
\title{Bootstrap Logistic Regression Confidence Intervals}
\usage{
bootstrap_logistic_confidence(obs, resp, pred, alpha = 0.05, B = 20)
}
\arguments{
\item{obs}{A \code{matrix} containing the full data set with predictors and response.}

\item{resp}{A \code{integer} representing the response variable column.}

\item{pred}{A \code{vector} representing the predictor variables.}

\item{alpha}{A \code{numeric} representing the significance level for the confidence intervals (default is 0.05).}

\item{B}{A \code{integer} representing the number of bootstrap samples to generate (default is 20).}
}
\value{
A \code{matrix} with coefficients and their corresponding lower and upper confidence intervals
}
\description{
Computes bootstrap confidence intervals for coefficients estimated by the logistic regression model.
}
\examples{
  #data obtained from https://stats.oarc.ucla.edu/r/dae/logit-regression/
  test_data <- read.csv("https://stats.idre.ucla.edu/stat/data/binary.csv")
  test_resp <- 1
  test_pred <- 2:4
  results <- suppressWarnings(bootstrap_logistic_confidence(test_data, test_resp, test_pred))
  print(results)
}
\author{
Group 11
}
