% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic_plot.R
\name{logistic_plot}
\alias{logistic_plot}
\title{Bootstrap Logistic Regression Confidence Intervals}
\usage{
logistic_plot(Data, Pred, Resp)
}
\arguments{
\item{Data}{A \code{matrix} containing the data we want to plot}

\item{Resp}{A \code{integer} representing the response variable column.}

\item{Beta}{A \code{vector} of beta values for the logistic regression (can be obtained from the \code{log_betas} function.)}

\item{pred}{A \code{integer} representing the predictor variable column.}
}
\value{
A plot showing the logistic regression line.
}
\description{
Computes bootstrap confidence intervals for coefficients estimated by the logistic regression model.
}
\examples{
  test_data <- read.csv("https://stats.idre.ucla.edu/stat/data/binary.csv")
  logistic_plot(test_data, Resp = 1, Pred= 3)
}
\author{
Group 11
}
