% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/part 4 plot of the confusion metrics.R
\name{plot_confusion_matrix}
\alias{plot_confusion_matrix}
\title{plot for “Confusion Matrix”}
\usage{
plot_confusion_matrix(Data, y, X, value_name = "prevalence")
}
\arguments{
\item{y}{: A  \code{integer} denoting the column of \code{Data} representing the response variable.}

\item{value_name}{: A character, representing the user's choice of which value in the matrix user wants to plot. It needs to be one of six `prevalence`, `accuracy`, `sensitivity`, `specificity`, `false_discovery_rate`, `diagnostic_odds_ratio` , the default is "prevalence"}

\item{Data:}{A  \code{matrix} containing the data we want to perform logistic regression on.}

\item{X:}{A  \code{vector} denoting the columns of \code{Data} representing the predictor variables.}
}
\value{
A plot of the value chosen from user versus cut_off_value(from0.1 to 0.9 with step of 0.1)
}
\description{
The possibility for the user to plot of any of the above metrics evaluated over a grid of cut-off values for prediction going from 0.1 to 0.9 with steps of 0.1.
}
\examples{
#data obtained from https://stats.oarc.ucla.edu/r/dae/logit-regression/
test_data <- read.csv("https://stats.idre.ucla.edu/stat/data/binary.csv")
  test_resp <- 1
  test_pred <- 2:4
result <- confusion_matrix(Data = test_data, X= test_pred, y = test_resp, cut_off = 0.5)
plot_confusion_matrix(Data = test_data, y = test_resp,X = test_pred , value_name = 'false_discovery_rate')
}
\author{
Group 11
}
